requirements.txt
openai
langchain
python-dotenv
unstructured
tqdm
tiktoken
chromadb
pysqlite3
pysqlite3-binary
jq
pinecone-client